RAG_script/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
dataclasses-json==0.6.7
distro==1.9.0
environs==9.5.0
exceptiongroup==1.2.2
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.18
langchain-experimental==0.0.62
langchain-milvus==0.1.1
langchain-openai==0.1.16
langchain-text-splitters==0.2.2
langsmith==0.1.85
marshmallow==3.21.3
milvus-lite==2.4.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.35.13
orjson==3.10.6
packaging==24.1
pandas==2.2.2
protobuf==5.27.2
pydantic==2.8.2
pydantic_core==2.20.1
pymilvus==2.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rank-bm25==0.2.2
regex==2024.5.15
requests==2.32.3
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
yarl==1.9.4